+++ fastapi_mvc/__main__.py
@@ -1,0 +1,6 @@
+"""fastapi-mvc."""
+from fastapi_mvc.cli.cli import cli
+
+
+if __name__ == "__main__":
+    cli()