+++ fastapi_mvc/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/app/exceptions/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""{{cookiecutter.project_description}}"""
+from .http import HTTPException, http_exception_handler
+
+
+__all__ = (HTTPException, http_exception_handler)