+++ fastapi_mvc/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/app/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""{{cookiecutter.project_description}}"""
+from .ready import ReadyResponse
+from .response import ErrorResponse
+
+__all__ = (
+    ReadyResponse,
+    ErrorResponse,
+)